pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "oneprot"
version = "0.0.1"
description = "Protein structure analysis and machine learning project"
readme = "README.md"
requires-python = ">=3.9,<3.13"
license = {text = "MIT"}
authors = [
    {name = "Klemens Floege", email = "your-email@example.com"},
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]

dependencies = [
    "pytorch-lightning==2.0.9",
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "torchmetrics>=0.11.0",
    "h5py==3.9.0",
    "ipykernel==6.23.1",
    "ipywidgets==8.1.1",
    "hydra-core==1.3.2",
    "hydra-colorlog==1.2.0",
    "hydra-optuna-sweeper",
    "biopython==1.81",
    "transformers==4.33.2",
    "peft==0.5.0",
    "fair-esm==2.0.0",
    "rootutils==1.0.7",
    "pyrootutils==1.0.4",
    "tqdm==4.65.0",
    "rich==13.5.3",
    # "dive-into-graphs",  # Temporarily commented due to Python 3.13 incompatibility
    "wandb",
    "numpy",
    "pandas",
    "scikit-learn",
    "matplotlib",
    "seaborn",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov",
    "pre-commit>=3.0.0",
    "black",
    "isort",
    "flake8",
    "mypy",
]
notebooks = [
    "jupyter",
    "jupyterlab",
    "notebook",
]

[project.scripts]
train_command = "oneprot.scripts.train:main"
eval_command = "oneprot.scripts.eval:main"

[project.urls]
Homepage = "https://github.com/klemensfloege/oneprot"
Repository = "https://github.com/klemensfloege/oneprot"
Issues = "https://github.com/klemensfloege/oneprot/issues"

[tool.setuptools.packages.find]
where = ["src"]
include = ["oneprot*"]

[tool.pytest.ini_options]
addopts = [
  "--color=yes",
  "--durations=0",
  "--strict-markers",
  "--doctest-modules",
]
filterwarnings = [
  "ignore::DeprecationWarning",
  "ignore::UserWarning",
]
log_cli = "True"
markers = [
  "slow: slow tests",
]
minversion = "6.0"
testpaths = "tests/"

[tool.coverage.report]
exclude_lines = [
    "pragma: nocover",
    "raise NotImplementedError",
    "raise NotImplementedError()",
    "if __name__ == .__main__.:",
]

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["src"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true